api/requirements.txt
fastapi
uvicorn[standard]
pyswisseph
pandas
